task2/requirements.txt
annotated-types==0.6.0
beaupy==3.8.2
black==24.4.2
click==8.1.7
dnspython==2.6.1
emoji==2.11.1
isort==5.13.2
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.1
python-dotenv==1.0.1
python-yakh==0.3.2
questo==0.2.3
rich==13.7.1
typing_extensions==4.11.0
